replacing model.py in starting kit
