+++ Starting_Kits/HEP/sample_code_submission/bootstrap.py
@@ -7,25 +7,40 @@ import json
 root_dir = "./"
 input_dir = root_dir + "input_data"
 
+def shuffle_dataframe(df):
+    # Convert DataFrame to NumPy array
+    array = df.to_numpy()
+
+    # Shuffle the array
+    np.random.shuffle(array)
+
+    # Convert the shuffled array back to a DataFrame
+    shuffled_df = pd.DataFrame(array, columns=df.columns)
+
+    return shuffled_df
+
 
 def bootstrap(weights, seed=42):
-
+    
     prng = RandomState(seed)
     n_obs = len(weights)
     new_weights = prng.poisson(lam=weights)
     return new_weights
 
+def bootstrap_data(data,weights,label,seed=42):
 
-def bootstrap_data(data, weights, label, n=1000, seed=42):
-
-    prng = RandomState(seed)
-    n_obs = len(weights)
-    new_weights = prng.poisson(lam=weights)
-
-    data['weights'] = new_weights
+    data['weights'] = weights
     data['label'] = label
-    data_bootstrap = data.sample(n)
-    return data_bootstrap
+    data_bootstrap = shuffle_dataframe(data)
+    del data
+    prng = RandomState(seed)
+
+    data_bootstrap['weights'] = prng.poisson(lam=data_bootstrap['weights'])
+
+    weights_bootstrap = data_bootstrap.pop('weights')
+    label_bootstrap = data_bootstrap.pop('label')
+
+    return {'data':data_bootstrap, 'weights':weights_bootstrap, 'label':label_bootstrap}
 
 
 if __name__ == '__main__':
@@ -46,10 +61,11 @@ if __name__ == '__main__':
     # read train settings
     with open(train_settings_file) as f:
         train_settings = json.load(f)
-
+    
     # read train weights
     with open(train_weights_file) as f:
         train_weights = np.array(f.read().splitlines(), dtype=float)
+
 
     train_set = {
         "data": train_data,
@@ -66,3 +82,9 @@ if __name__ == '__main__':
 
     total_weights_distribution_array = np.array(total_weights_distribution)
     np.histogram(total_weights_distribution_array, bins=100)
+
+
+
+
+
+
