+++ Starting_Kits/HEP/sample_code_submission/model.py
@@ -3,11 +3,12 @@ from sys import path
 import numpy as np
 import pandas as pd
 from math import sqrt, log
+from xgboost import XGBClassifier
 from xgboost import XGBRegressor
 from sklearn.utils import shuffle
+from sklearn.metrics import roc_auc_score
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler
-from sklearn.multioutput import MultiOutputRegressor
 
 # ------------------------------
 # Absolute path to submission dir
@@ -17,7 +18,6 @@ path.append(submissions_dir)
 
 from bootstrap import *
 from systematics import postprocess
-
 # ------------------------------
 # Constants
 # ------------------------------
@@ -69,9 +69,9 @@ class Model():
         self.systematics = systematics
 
         # Intialize class variables
-        self.validation_sets = None
-        self.theta_candidates = np.arange(0.5, 0.99, 0.01)
-        self.best_theta = 0.9
+        self.validation_set = None
+        self.theta_candidates = np.linspace(0.8, 1.0, 100)
+        self.best_theta = 0.8
         self.scaler = StandardScaler()
         self.scaler_tes = StandardScaler()
 
@@ -91,7 +91,7 @@ class Model():
         self._init_model()
         self._train()
         self._choose_theta()
-        self.mu_hat_calc()
+        # self.mu_hat_calc()
         self._validate()
         self._compute_validation_result()
 
@@ -111,24 +111,28 @@ class Model():
                 - p84
         """
 
+        print("[*] - Testing")
         test_df = test_set['data']
-        test_df = self.scaler.transform(test_df)
         Y_hat_test = self._predict(test_df, self.best_theta)
 
+        print("[*] - Computing Test result")
         weights_train = self.train_set["weights"].copy()
         weights_test = test_set["weights"].copy()
 
+        print(f"[*] --- total weight test: {weights_test.sum()}") 
+        print(f"[*] --- total weight train: {weights_train.sum()}")
+        print(f"[*] --- total weight mu_cals_set: {self.mu_calc_set['weights'].sum()}")
+
+        weight = weights_test*(Y_hat_test)
         # get n_roi
-        n_roi = self.N_calc_2(weights_test[Y_hat_test == 1])
 
-        mu_hat = (n_roi - self.beta_roi)/self.gamma_roi
+        mu_hat,mu_p16,mu_p84 = self._compute_result(weight)
+        delta_mu_hat = np.abs(mu_p84 - mu_p16)
 
-        sigma_mu_hat = np.std(mu_hat)
-
-        delta_mu_hat = 2*sigma_mu_hat
-
-        mu_p16 = np.percentile(mu_hat, 16)
-        mu_p84 = np.percentile(mu_hat, 84)
+        print(f"[*] --- mu_hat: {mu_hat.mean()}")
+        print(f"[*] --- delta_mu_hat: {delta_mu_hat}")
+        print(f"[*] --- p16: {mu_p16}")
+        print(f"[*] --- p84: {mu_p84}")
 
         return {
             "mu_hat": mu_hat.mean(),
@@ -138,11 +142,13 @@ class Model():
         }
 
     def _init_model(self):
-        print("[*] - Intialize Baseline Model (XGBoostRegressor)")
-
-        xgb = XGBRegressor(learning_rate=0.05, n_estimators=100)
-        self.model = MultiOutputRegressor(xgb)
-
+        print("[*] - Intialize Baseline Model (XBM Classifier Model)")
+        self.model_tes = XGBRegressor()
+        self.model = XGBClassifier(
+            tree_method="hist",
+            use_label_encoder=False,
+            eval_metric=['logloss', 'auc']
+        )
     def _generate_validation_sets(self):
         print("[*] - Generating Validation sets")
 
@@ -155,7 +161,7 @@ class Model():
             self.train_set["data"],
             self.train_set["labels"],
             self.train_set["weights"],
-            test_size=0.05,
+            test_size=0.2,
             stratify=self.train_set["labels"]
         )
 
@@ -167,6 +173,7 @@ class Model():
             shuffle=True,
             stratify=train_labels
         )
+
 
         # Calculate the sum of weights for signal and background in the training and validation sets
         train_signal_weights = train_weights[train_labels == 1].sum()
@@ -184,17 +191,6 @@ class Model():
         mu_calc_set_weights[mu_calc_set_labels == 1] *= signal_weights / mu_calc_set_signal_weights
         mu_calc_set_weights[mu_calc_set_labels == 0] *= background_weights / mu_calc_set_background_weights
 
-        self.train_df = train_df
-
-        self.train_set = {
-            "data": train_df,
-            "labels": train_labels,
-            "weights": train_weights,
-            "settings": self.train_set["settings"]
-        }
-
-        self.eval_set = [(self.train_set['data'], self.train_set['labels']), (valid_df.to_numpy(), valid_labels)]
-        
         train_df = train_df.copy()
         train_df["weights"] = train_weights
         train_df["labels"] = train_labels
@@ -207,11 +203,34 @@ class Model():
         mu_calc_set_df = mu_calc_set_df.copy()
         mu_calc_set_df["weights"] = mu_calc_set_weights
         mu_calc_set_df["labels"] = mu_calc_set_labels
+
         mu_calc_set_df = postprocess(mu_calc_set_df)
 
         mu_calc_set_weights = mu_calc_set_df.pop('weights')
         mu_calc_set_labels = mu_calc_set_df.pop('labels')
 
+        # valid_df = valid_df.copy()
+        # valid_df["weights"] = valid_weights
+        # valid_df["labels"] = valid_labels
+
+        # valid_df = postprocess(valid_df)
+
+        # valid_weights = valid_df.pop('weights')
+        # valid_labels = valid_df.pop('labels')
+
+
+
+
+        self.train_df = train_df
+
+        self.train_set = {
+            "data": train_df,
+            "labels": train_labels,
+            "weights": train_weights,
+            "settings": self.train_set["settings"]
+        }
+
+        self.eval_set = [(self.train_set['data'], self.train_set['labels']), (valid_df.to_numpy(), valid_labels)]
 
         self.mu_calc_set = {
                 "data": mu_calc_set_df,
@@ -220,18 +239,19 @@ class Model():
             }
 
         self.validation_sets = []
-        # Loop 10 times to generate 10 validation sets
-        for i in range(0, 20):
+        for i in range(10):
+            # Loop 10 times to generate 10 validation sets
             tes = round(np.random.uniform(0.9, 1.10), 2)
             # apply systematics
             valid_df_temp = valid_df.copy()
             valid_df_temp["weights"] = valid_weights
             valid_df_temp["labels"] = valid_labels
 
-            valid_with_systematics_temp = self.systematics(
-                data=valid_df_temp,
-                tes=tes
-            ).data
+            # valid_with_systematics_temp = self.systematics(
+            #     data=valid_df_temp,
+            #     tes=tes
+            # ).data
+            valid_with_systematics_temp = postprocess(valid_df_temp)
 
             valid_labels_temp = valid_with_systematics_temp.pop('labels')
             valid_weights_temp = valid_with_systematics_temp.pop('weights')
@@ -247,6 +267,7 @@ class Model():
             del valid_with_systematics_temp
             del valid_df_temp
 
+        self.validation_set = self.validation_sets[1]
 
         train_signal_weights = train_weights[train_labels == 1].sum()
         train_background_weights = train_weights[train_labels == 0].sum()
@@ -260,15 +281,20 @@ class Model():
         print(f"[*] --- valid signal: {valid_signal_weights} --- valid background: {valid_background_weights}")
         print(f"[*] --- mu_calc_set signal: {mu_calc_set_signal_weights} --- mu_calc_set background: {mu_calc_set_background_weights}")
 
-    def _train(self):
+    def _train_2(self):
 
         tes_sets = []
 
-        for i in range(0, 10):
-            tes_set = self.train_df.copy()
+        for i in range(0, 100):
+
+            tes_set = self.train_set['data'].copy()
+
             tes_set = pd.DataFrame(tes_set)
+
             tes_set["weights"] = self.train_set["weights"]
             tes_set["labels"] = self.train_set["labels"]
+
+            tes_set = tes_set.sample(frac=0.01, replace=True, random_state=i).reset_index(drop=True)
 
             # adding systematics to the tes set
             # Extract the TES information from the JSON file
@@ -284,7 +310,7 @@ class Model():
 
             data_syst = data_syst.round(3)
             tes_set = data_syst.copy()
-            tes_set['tes'] = (tes*10)*2
+            tes_set['tes'] = tes
             tes_sets.append(tes_set)
             del data_syst
             del tes_set
@@ -293,10 +319,9 @@ class Model():
 
         train_tes_data = shuffle(tes_sets_df)
 
-        tes_label_1 = train_tes_data.pop('tes')
+        tes_label = train_tes_data.pop('tes')
         tes_label_2 = train_tes_data.pop('labels')
-        tes_label = [tes_label_1, tes_label_2]
-        tes_label = np.array(tes_label).T
+
         tes_weights = train_tes_data.pop('weights')
 
         weights_train = tes_weights.copy()
@@ -311,56 +336,123 @@ class Model():
         print("[*] --- Training Model")
         train_tes_data = self.scaler.fit_transform(train_tes_data)
 
-        self._fit(train_tes_data, tes_label, weights_train)
+        print("[*] --- shape of train tes data", train_tes_data.shape)
+        
+        self.model_tes.fit(train_tes_data, tes_label, sample_weight=tes_weights)
+
+
+    def _train(self):
+
+        self._train_2()
+        weights_train = self.train_set["weights"].copy()
+        train_labels = self.train_set["labels"].copy()
+        train_data = self.train_set["data"].copy()
+        class_weights_train = (weights_train[train_labels == 0].sum(), weights_train[train_labels == 1].sum())
+
+        for i in range(len(class_weights_train)):  # loop on B then S target
+            # training dataset: equalize number of background and signal
+            weights_train[train_labels == i] *= max(class_weights_train) / class_weights_train[i]
+            # test dataset : increase test weight to compensate for sampling
+
+        print("[*] --- Training Model")
+
+        print("[*] --- shape of train tes data", train_data.shape)
+
+        self._fit(train_data, train_labels, weights_train)
 
         print("[*] --- Predicting Train set")
         self.train_set['predictions'] = (self.train_set['data'], self.best_theta)
 
         self.train_set['score'] = self._return_score(self.train_set['data'])
 
+        auc_train = roc_auc_score(
+            y_true=self.train_set['labels'],
+            y_score=self.train_set['score'],
+            sample_weight=self.train_set['weights']
+        )
+        print(f"[*] --- AUC train : {auc_train}")
+
     def _fit(self, X, y, w):
         print("[*] --- Fitting Model")
-        self.model.fit(X, y, sample_weight=w)
+                
+        tes = self.model_tes.predict(X)
+        X_df = pd.DataFrame(X)
+        X_df['tes'] = tes
+        X_ = X_df.to_numpy()
+        X_ = self.scaler.fit_transform(X_)
+        self.model.fit(X_, y, sample_weight=w)
 
     def _return_score(self, X):
-        y_predict = self.model.predict(X)[:, 1]
+        tes = self.model_tes.predict(X)
+        X_df = pd.DataFrame(X)
+        X_df['tes'] = tes
+        X_ = X_df.to_numpy()
+        X_ = self.scaler.fit_transform(X_)
+        y_predict = self.model.predict(X_)
         return y_predict
 
     def _predict(self, X, theta):
         Y_predict = self._return_score(X)
-        predictions = np.where(Y_predict > theta, 1, 0)
+        predictions = (Y_predict > theta).astype(int)
         return predictions
 
-    def N_calc_2(self, weights, n=100):
+    def N_calc_2(self, weights, n=1000):
         total_weights = []
         for i in range(n):
             bootstrap_weights = bootstrap(weights=weights, seed=42+i)
-            bootstrap_weights_array = np.array(bootstrap_weights)
-            effective_number = (bootstrap_weights_array).sum()
-            total_weights.append(effective_number)
+            total_weights.append(np.array(bootstrap_weights).sum())
         n_calc_array = np.array(total_weights)
         return n_calc_array
 
-    def mu_hat_calc(self):
 
-        self.mu_calc_set['data'] = self.scaler.transform(self.mu_calc_set['data'])
-        Y_hat_mu_calc_set = self._predict(self.mu_calc_set['data'], self.best_theta)
-        Y_mu_calc_set = self.mu_calc_set['labels']
-        weights_mu_calc_set = self.mu_calc_set['weights']
+    # def mu_hat_calc(self):  
 
-        # compute gamma_roi
-        weights_mu_calc_set_signal = weights_mu_calc_set[Y_mu_calc_set == 1]
-        weights_mu_calc_set_bkg = weights_mu_calc_set[Y_mu_calc_set == 0]
+    #     self.mu_calc_set['data'] = self.scaler.transform(self.mu_calc_set['data'])
+    #     Y_hat_mu_calc_set = self._predict(self.mu_calc_set['data'], self.best_theta)
+    #     Y_mu_calc_set = self.mu_calc_set['labels']
+    #     weights_mu_calc_set = self.mu_calc_set['weights']
 
-        Y_hat_mu_calc_set_signal = Y_hat_mu_calc_set[Y_mu_calc_set == 1]
-        Y_hat_mu_calc_set_bkg = Y_hat_mu_calc_set[Y_mu_calc_set == 0]
+    #     # compute gamma_roi
+    #     weights_mu_calc_set_signal = weights_mu_calc_set[Y_mu_calc_set == 1]
+    #     weights_mu_calc_set_bkg = weights_mu_calc_set[Y_mu_calc_set == 0]
 
-        self.gamma_roi = (weights_mu_calc_set_signal[Y_hat_mu_calc_set_signal == 1]).sum()
+    #     Y_hat_mu_calc_set_signal = Y_hat_mu_calc_set[Y_mu_calc_set == 1]
+    #     Y_hat_mu_calc_set_bkg = Y_hat_mu_calc_set[Y_mu_calc_set == 0]
 
-        # compute beta_roi
-        self.beta_roi = (weights_mu_calc_set_bkg[Y_hat_mu_calc_set_bkg == 1]).sum()
-        if self.gamma_roi == 0:
-            self.gamma_roi = EPSILON
+    #     self.gamma_roi = (weights_mu_calc_set_signal[Y_hat_mu_calc_set_signal == 1]).sum()
+
+    #     # compute beta_roi
+    #     self.beta_roi = (weights_mu_calc_set_bkg[Y_hat_mu_calc_set_bkg == 1]).sum()
+    #     if self.gamma_roi == 0:
+    #         self.gamma_roi = EPSILON
+
+    def amsasimov_x(self, s, b):
+        '''
+        This function calculates the Asimov crossection significance for a given number of signal and background events.
+        Parameters: s (float) - number of signal events
+
+        Returns:    float - Asimov crossection significance
+        '''
+
+        if b <= 0 or s <= 0:
+            return 0
+        try:
+            return s/sqrt(s+b)
+        except ValueError:
+            print(1+float(s)/b)
+            print(2*((s+b)*log(1+float(s)/b)-s))
+        # return s/sqrt(s+b)
+
+    def del_mu_stat(self, s, b):
+        '''
+        This function calculates the statistical uncertainty on the signal strength.
+        Parameters: s (float) - number of signal events
+                    b (float) - number of background events
+
+        Returns:    float - statistical uncertainty on the signal strength
+
+        '''
+        return (np.sqrt(s + b)/s)
 
     def get_meta_validation_set(self):
 
@@ -382,106 +474,172 @@ class Model():
     def _choose_theta(self):
 
         print("[*] Choose best theta")
-        meta_validation_set = self.get_meta_validation_set()
-        theta_sigma_squared = []
 
+        meta_validation_set = self.validation_set
+        val_min = 1
         # Loop over theta candidates
         # try each theta on meta-validation set
         # choose best theta
         for theta in self.theta_candidates:
-            meta_validation_set_df = self.scaler.transform(meta_validation_set["data"])    
+            meta_validation_set_df = (meta_validation_set["data"])    
             # Get predictions from trained model
-            Y_hat_valid = self._predict(meta_validation_set_df, theta)
-            Y_valid = meta_validation_set["labels"]
 
-            weights_valid = meta_validation_set["weights"].copy()
 
             # get region of interest
-            nu_roi = (weights_valid[Y_hat_valid == 1]).sum()/10
 
-            weights_valid_signal = weights_valid[Y_valid == 1]  
-            weights_valid_bkg = weights_valid[Y_valid == 0]
+            # predict probabilities for holdout
+            X_holdout_sc = (self.mu_calc_set['data'])
+            w_holdout = self.mu_calc_set['weights']
+            y_holdout = self.mu_calc_set['labels']
+            y_pred = self._predict(X_holdout_sc, theta)
+            
 
-            Y_hat_valid_signal = Y_hat_valid[Y_valid == 1]  
-            Y_hat_valid_bkg = Y_hat_valid[Y_valid == 0] 
+            gamma_roi = (w_holdout*(y_pred * y_holdout)).sum()
+            beta_roi = (w_holdout*(y_pred * (1-y_holdout))).sum()
 
-            # compute gamma_roi
-            gamma_roi = (weights_valid_signal[Y_hat_valid_signal == 1]).sum()/10
 
-            # compute beta_roi
-            beta_roi = (weights_valid_bkg[Y_hat_valid_bkg == 1]).sum()/10
+            Y_hat_valid = self._predict(meta_validation_set_df, theta)
+            weights_valid = meta_validation_set["weights"].copy() 
 
-            # Compute sigma squared mu hat
-            sigma_squared_mu_hat = nu_roi/np.square(gamma_roi)
+            weight = weights_valid*(Y_hat_valid)
 
-            # get N_ROI from predictions
-            theta_sigma_squared.append(sigma_squared_mu_hat)
+            mu_scan = np.linspace(0, 3, 100)
+            hist_llr = self.calculate_NLL(mu_scan, weight,gamma_roi,beta_roi)
+            hist_llr = np.array(hist_llr)
 
-            print(f"\n[*] --- theta: {theta}--- nu_roi: {nu_roi} --- beta_roi: {beta_roi} --- gamma_roi: {gamma_roi} --- sigma squared: {sigma_squared_mu_hat}")
+            val =  np.abs(mu_scan[np.argmin(hist_llr)] - 1)
 
-        # Choose theta with min sigma squared
-        try:
-            index_of_least_sigma_squared = np.nanargmin(theta_sigma_squared)
-        except:
-            print("[!] - WARNING! All sigma squared are nan")
-            index_of_least_sigma_squared = np.argmin(theta_sigma_squared)
+            if val < val_min:
+                val_min = val
+                print("val: ", val)
+                print("gamma_roi: ", gamma_roi)
+                print("beta_roi: ", beta_roi)
+                print("Uncertainity", np.sqrt(gamma_roi + beta_roi)/gamma_roi)
+                Beta_roi = beta_roi.copy()
+                Gamma_roi = gamma_roi.copy()
+                self.best_theta = theta
 
-        self.best_theta = self.theta_candidates[index_of_least_sigma_squared]
+
+        theta = self.best_theta
+        # predict probabilities for holdout
+        X_holdout_sc = (self.mu_calc_set['data'])
+        w_holdout = self.mu_calc_set['weights']
+        y_holdout = self.mu_calc_set['labels']
+        y_pred = self._predict(X_holdout_sc, theta)
+        
+
+        gamma_roi = (w_holdout*(y_pred * y_holdout)).sum()
+        beta_roi = (w_holdout*(y_pred * (1-y_holdout))).sum()
+
+
+        Y_hat_valid = self._predict(meta_validation_set_df, theta)
+        weights_valid = meta_validation_set["weights"].copy() 
+
+        weight = weights_valid*(Y_hat_valid)
+
+        mu_scan = np.linspace(0, 3, 100)
+        hist_llr = self.calculate_NLL(mu_scan, weight,gamma_roi,beta_roi)
+        hist_llr = np.array(hist_llr)
+
+        val =  np.abs(mu_scan[np.argmin(hist_llr)] - 1)
+        print("val: ", val)
+        print("gamma_roi: ", gamma_roi)
+        print("beta_roi: ", beta_roi)
+        print("Uncertainity", np.sqrt(gamma_roi + beta_roi)/gamma_roi)
+        self.beta_roi = beta_roi.copy()
+        self.gamma_roi = gamma_roi.copy()
+
+        self.force_correction_term = val
         print(f"[*] --- Best theta : {self.best_theta}")
 
     def _validate(self):
         for valid_set in self.validation_sets:
-            valid_set['data'] = self.scaler.transform(valid_set['data'])
             valid_set['predictions'] = self._predict(valid_set['data'], self.best_theta)
             valid_set['score'] = self._return_score(valid_set['data'])
+
+
+    def calculate_NLL( self,mu_scan, weight_data,gamma_roi,beta_roi):
+        def _sigma_asimov_SR(mu):
+            return mu*gamma_roi + beta_roi
+
+        sum_data_total_SR = weight_data.sum()
+        comb_llr = []
+        for i, mu in enumerate(mu_scan):
+            hist_llr = (
+                -2
+                * sum_data_total_SR
+                * np.log((_sigma_asimov_SR(mu) / _sigma_asimov_SR(1.0)))
+            ) + (2 * (_sigma_asimov_SR(mu) - _sigma_asimov_SR(1.0)))
+
+            comb_llr.append(hist_llr )
+
+        comb_llr = np.array(comb_llr)
+        comb_llr = comb_llr - np.amin(comb_llr)
+
+        return comb_llr
+
+    
+    def _sigma_asimov_SR(self,mu):
+        return mu*self.gamma_roi + self.beta_roi
+
+    def _compute_result(self,weights):
+        mu_scan = np.linspace(0, 5, 100)
+        sum_data_total_SR = weights.sum()
+        comb_llr = []
+        for i, mu in enumerate(mu_scan):
+            hist_llr = (
+                -2
+                * sum_data_total_SR
+                * np.log((self._sigma_asimov_SR(mu) / self._sigma_asimov_SR(1.0)))
+            ) + (2 * (self._sigma_asimov_SR(mu) - self._sigma_asimov_SR(1.0)))
+            comb_llr.append(hist_llr)
+
+        comb_llr = np.array(comb_llr)
+        comb_llr = comb_llr - np.amin(comb_llr)
+
+
+        if (mu_scan[np.where((comb_llr <= 9.0) & (comb_llr >= 0.0))].size == 0):
+            p16 = 0
+            p84 = 0
+            mu = 0
+        else:
+            p16 = min(mu_scan[np.where((comb_llr <= 9.0) & (comb_llr >= 0.0))])
+            p84 = max(mu_scan[np.where((comb_llr <= 9.0) & (comb_llr >= 0.0))]) 
+            mu = mu_scan[np.argmin(comb_llr)]
+
+        mu = mu - self.force_correction_term
+        p16 = p16 - self.force_correction_term
+        p84 = p84 - self.force_correction_term
+        return mu, p16, p84
 
     def _compute_validation_result(self):
         print("[*] - Computing Validation result")
 
         self.validation_delta_mu_hats = []
         for valid_set in self.validation_sets:
-            Y_hat_train = self.train_set["predictions"]
-            Y_train = self.train_set["labels"]
             Y_hat_valid = valid_set["predictions"]
-            Y_valid = valid_set["labels"]
-            Score_train = self.train_set["score"]
             Score_valid = valid_set["score"]
+
+            auc_valid = roc_auc_score(y_true=valid_set["labels"], y_score=Score_valid,sample_weight=valid_set['weights'])
+            print(f"\n[*] --- AUC validation : {auc_valid} --- tes : {valid_set['tes']}")
 
             # print(f"[*] --- PRI_had_pt : {valid_set['had_pt']}")
             # del Score_valid
-            weights_train = self.train_set["weights"].copy()
             weights_valid = valid_set["weights"].copy()
 
-            signal_valid = weights_valid[Y_valid == 1]
-            background_valid = weights_valid[Y_valid == 0]
+            weight = weights_valid*(Y_hat_valid)
 
-            Y_hat_valid_signal = Y_hat_valid[Y_valid == 1]
-            Y_hat_valid_bkg = Y_hat_valid[Y_valid == 0]
+            mu_hat,mu_p16,mu_p84 = self._compute_result(weight)
 
-            signal = signal_valid[Y_hat_valid_signal == 1].sum()
-            background = background_valid[Y_hat_valid_bkg == 1].sum()
 
-            # get n_roi
-            n_roi = self.N_calc_2(weights_valid[Y_hat_valid == 1])
-
-            mu_hat = ((n_roi - self.beta_roi)/self.gamma_roi).mean()
-            # get region of interest
-            nu_roi = self.beta_roi + self.gamma_roi
-
-            gamma_roi = self.gamma_roi
-
-            # compute beta_roi
-            beta_roi = self.beta_roi
-            if gamma_roi == 0:
-                gamma_roi = EPSILON
-
-            # Compute mu_hat
 
             # Compute delta mu hat (absolute value)
             delta_mu_hat = np.abs(valid_set["settings"]["ground_truth_mu"] - mu_hat)
 
             self.validation_delta_mu_hats.append(delta_mu_hat)
 
+
             print(f"[*] --- mu: {np.round(valid_set['settings']['ground_truth_mu'], 4)} --- mu_hat: {np.round(mu_hat, 4)} --- delta_mu_hat: {np.round(delta_mu_hat, 4)}")
+            print(f"[*] --- p16: {np.round(mu_p16, 4)} --- p84: {np.round(mu_p84, 4)} --- mu_hat_: {np.round(mu_hat, 4)}")
 
         print(f"[*] --- validation delta_mu_hat (avg): {np.round(np.mean(self.validation_delta_mu_hats), 4)}")
